Import www.brettchalupa.com Jekyll setup
This overwrites most of the site to make use of the setup from my
personal website
(https://github.com/brettchalupa/brettchalupa.github.io). This commit
includes the clearing out of files that are not needed, as well as some
minor adjustments to make it unique for Monoso. From here, I'll adjust
the styles and content to be Monoso specific.