Add services & projects
This is the first round of adding content to the site to populate the
projects and services. It clears out the previous placeholder content.
Most of the entries need to be expanded upon, but it gets it set up.
This also adds a page for listing out the newly added services.