Add Disqus comment support
Posts need to have `comments: true` in the front-matter to support
comments. This will, hopefully, allow readers to ask questions and share
feedback on blog posts and screencasts.

I'm not 100% sold on Disqus, but it seems like it could be a good fit,
so I'm going to give it a try.