Initial adjustments for portfolio transition
I'm making Monoso my portfolio for my technical projects so that it is
less of an ominous business entity and more personal. It'll be the place
to house my technical projects, screencasts, and blog posts.

This commit gets the ball rolling by updating the home and about
language to be first person, unfeaturing client projects, and tweaking
the sections of the home page to focus on the blog more.